Add YJIT build for rails CI
This change adds a build for CI that will run rails main merges on Ruby
master with YJIT enabled. This will enable us to test against YJIT and
catch any potential regressions early. This will also make it easier to
find optimizations in YJIT that will benefit Rails because we can be
confident Rails is runnable with YJIT which will make it easier to
collect stats.