Display debian codename when building image
This is a small quality of life improvement which could help make it
easier to debug issues in the future. There have been issues previously
when our base Ruby image changes its debian release under us and this
type of change can help make it more obvious when that happens.

Additionally, this should help us have better insight into when we can
change/remove codename specific workarounds (for example we currently
have a workaround for Debian 8 (jessie) in the Dockerfile but its not
obvious when that code is being run)