Remove pipeline_slug
This was initially designed to be a way to restrict usage to officially supported pipelines.

Since pipelines for buildkite-config and modified rails-ci builds require approval, this extra friction seems unnecessary and can actually introduce confusion when modifying or creating new pipelines.

For example, if we add a new pipeline we want to generate via the `pipeline-generate` script, we also need to add it to the list of approved pipelines.