Use a SELECT query for mysql healthcheck
Since updating the MySQL healthcheck to use mysqladmin, there has been
an increase in failing buildkite jobs that report being unable to
connect to MySQL. This behavior is consistent with reports on
stackoverflow that the mysqladmin check is unfortunately not a reliable
way to know whether the database is ready for connections.

This commit changes the MySQL healthcheck to be a `SELECT 1;` query.
This should ensure that the MySQL is not only started but additionally
ready for connections before reporting as healthy. Additionally, the
interval for the healthcheck was reduced to 1s so that it has the
potential to pass faster instead of only in 10s intervals.